+++ geopandas/plotting.py
@@ -314,10 +314,10 @@ def plot_series(s, cmap=None, color=None, ax=None, figsize=None, **style_kwds):
     return ax
 
 
-def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
+def plot_dataframe(df, column=None, cmap=None, color=None, ax=None, cax=None,
                    categorical=False, legend=False, scheme=None, k=5,
                    vmin=None, vmax=None, markersize=None, figsize=None,
-                   legend_kwds=None, **style_kwds):
+                   legend_kwds=None, classification_kwds=None, **style_kwds):
     """
     Plot a GeoDataFrame.
 
@@ -342,6 +342,8 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
         If specified, all objects will be colored uniformly.
     ax : matplotlib.pyplot.Artist (default None)
         axes on which to draw the plot
+    cax : matplotlib.pyplot Artist (default None)
+        axes on which to draw the legend in case of color map.
     categorical : bool (default False)
         If False, cmap will reflect numerical values of the
         column being plotted.  For non-numerical columns, this
@@ -351,8 +353,12 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
     scheme : str (default None)
         Name of a choropleth classification scheme (requires mapclassify).
         A mapclassify.Map_Classifier object will be used
-        under the hood. Supported schemes: 'Quantiles',
-        'Equal_Interval', 'Fisher_Jenks', 'Fisher_Jenks_Sampled'
+        under the hood. Supported are all schemes provided by mapclassify (e.g.
+        'Box_Plot', 'Equal_Interval', 'Fisher_Jenks', 'Fisher_Jenks_Sampled',
+        'HeadTail_Breaks', 'Jenks_Caspall', 'Jenks_Caspall_Forced',
+        'Jenks_Caspall_Sampled', 'Max_P_Classifier', 'Maximum_Breaks',
+        'Natural_Breaks', 'Quantiles', 'Percentiles', 'Std_Mean',
+        'User_Defined'). Arguments can be passed in classification_kwds.
     k : int (default 5)
         Number of classes (ignored if scheme is None)
     vmin : None or float (default None)
@@ -372,6 +378,8 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
         axes is given explicitly, figsize is ignored.
     legend_kwds : dict (default None)
         Keyword arguments to pass to ax.legend()
+    classification_kwds : dict (default None)
+        Keyword arguments to pass to mapclassify
 
     **style_kwds : dict
         Color options to be passed on to the actual plot function, such
@@ -400,6 +408,8 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
     import matplotlib.pyplot as plt
 
     if ax is None:
+        if cax is not None:
+            raise ValueError("'ax' can not be None if 'cax' is not.")
         fig, ax = plt.subplots(figsize=figsize)
     ax.set_aspect('equal')
 
@@ -445,7 +455,12 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
         values = np.array([valuemap[k] for k in values])
 
     if scheme is not None:
-        binning = _mapclassify_choro(values, scheme, k=k)
+        if classification_kwds is None:
+            classification_kwds = {}
+        if 'k' not in classification_kwds:
+                classification_kwds['k'] = k
+
+        binning = _mapclassify_choro(values, scheme, **classification_kwds)
         # set categorical to True for creating the legend
         categorical = True
         binedges = [values.min()] + binning.bins.tolist()
@@ -453,8 +468,8 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
                       for i in range(len(binedges)-1)]
         values = np.array(binning.yb)
 
-    mn = values.min() if vmin is None else vmin
-    mx = values.max() if vmax is None else vmax
+    mn = values[~np.isnan(values)].min() if vmin is None else vmin
+    mx = values[~np.isnan(values)].max() if vmax is None else vmax
 
     geom_types = df.geometry.type
     poly_idx = np.asarray((geom_types == 'Polygon')
@@ -475,6 +490,11 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
     if not lines.empty:
         plot_linestring_collection(ax, lines, values[line_idx],
                                    vmin=mn, vmax=mx, cmap=cmap, **style_kwds)
+
+    if cax is not None:
+        cbar_kwargs = {"cax": cax}
+    else:
+        cbar_kwargs = {"ax": ax}
 
     # plot all Points in the same collection
     points = df.geometry[point_idx]
@@ -507,13 +527,13 @@ def plot_dataframe(df, column=None, cmap=None, color=None, ax=None,
             ax.legend(patches, categories, **legend_kwds)
         else:
             n_cmap.set_array([])
-            ax.get_figure().colorbar(n_cmap, ax=ax)
+            ax.get_figure().colorbar(n_cmap, **cbar_kwargs)
 
     plt.draw()
     return ax
 
 
-def _mapclassify_choro(values, scheme, k=5):
+def _mapclassify_choro(values, scheme, **classification_kwds):
     """
     Wrapper for choropleth schemes from mapclassify for use with plot_dataframe
 
@@ -523,10 +543,16 @@ def _mapclassify_choro(values, scheme, k=5):
         Series to be plotted
     scheme : str
         One of mapclassify classification schemes
-        Options are 'Quantiles', 'Equal_Interval', 'Fisher_Jenks',
-        'Fisher_Jenks_Sampled'
-    k : int
-        number of classes (2 <= k <=9)
+        Options are Box_Plot, Equal_Interval, Fisher_Jenks,
+        Fisher_Jenks_Sampled, HeadTail_Breaks, Jenks_Caspall,
+        Jenks_Caspall_Forced, Jenks_Caspall_Sampled, Max_P_Classifier,
+        Maximum_Breaks, Natural_Breaks, Quantiles, Percentiles, Std_Mean,
+        User_Defined
+
+    **classification_kwds : dict
+        Keyword arguments for classification scheme
+        For details see mapclassify documentation:
+        https://mapclassify.readthedocs.io/en/latest/api.html
 
     Returns
     -------
@@ -536,28 +562,31 @@ def _mapclassify_choro(values, scheme, k=5):
 
     """
     try:
-        from mapclassify import (
-            Quantiles, Equal_Interval, Fisher_Jenks,
-            Fisher_Jenks_Sampled)
+        import mapclassify.classifiers
     except ImportError:
-        try:
-            from mapclassify.api import (
-                Quantiles, Equal_Interval, Fisher_Jenks,
-                Fisher_Jenks_Sampled)
-        except ImportError:
-            raise ImportError(
-                "The 'mapclassify' package is required to use the 'scheme' "
-                "keyword")
+        raise ImportError(
+            "The 'mapclassify' package is required to use the 'scheme' "
+            "keyword")
 
     schemes = {}
-    schemes['equal_interval'] = Equal_Interval
-    schemes['quantiles'] = Quantiles
-    schemes['fisher_jenks'] = Fisher_Jenks
-    schemes['fisher_jenks_sampled'] = Fisher_Jenks_Sampled
+    for classifier in mapclassify.classifiers.CLASSIFIERS:
+        schemes[classifier.lower()] = getattr(mapclassify.classifiers,
+                                              classifier)
 
     scheme = scheme.lower()
     if scheme not in schemes:
         raise ValueError("Invalid scheme. Scheme must be in the"
                          " set: %r" % schemes.keys())
-    binning = schemes[scheme](values, k)
+    if classification_kwds['k'] is not None:
+        try:
+            from inspect import getfullargspec as getspec
+        except ImportError:
+            from inspect import getargspec as getspec
+        spec = getspec(schemes[scheme].__init__)
+        if 'k' not in spec.args:
+            del classification_kwds['k']
+    try:
+        binning = schemes[scheme](values, **classification_kwds)
+    except TypeError:
+        raise TypeError("Invalid keyword argument for %r " % scheme)
     return binning